requirements.txt
dj-database-url==2.3.0
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
uvicorn==0.30.6
psycopg2-binary==2.9.10
requests==2.31.0